Extract CreatePartyScreen
